pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coreui_django_component"
version = "0.0.1"
description = "Python Boilerplate contains all the boilerplate you need to create a Python package."
readme = "README.rst"
authors = [
  {name = "6F-TECH-CONSULTANT-INDIA-Pvt-Ltd", email = "support@a6ftechconsultant.com"}
]
maintainers = [
  {name = "6F-TECH-CONSULTANT-INDIA-Pvt-Ltd", email = "support@a6ftechconsultant.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
    "django-viewcomponent",
    "django-compressor"
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

bugs = "https://github.com/6F-TECH-CONSULTANT-INDIA-Pvt-Ltd/coreui_django_component/issues"
changelog = "https://github.com/6F-TECH-CONSULTANT-INDIA-Pvt-Ltd/coreui_django_component/blob/master/changelog.md"
homepage = "https://github.com/6F-TECH-CONSULTANT-INDIA-Pvt-Ltd/coreui_django_component"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

